backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
crewai>=0.28.0
crewai-tools>=0.4.0
